feat: updated function getCommandPageData
