+++ src/scraper/frameDataPage/tableScrape.mjs
@@ -1,4 +1,4 @@
-import { handleImage, replaceWithAbbreviation } from '../helperFunctions/handleImage.mjs'
+import { replaceWithAbbreviation } from '../helperFunctions/handleImage.mjs'
 
 function extractDataFromTable() {
 
@@ -42,33 +42,33 @@ function extractDataFromTable() {
   
   let styleIndex = -1;
 
-  const getText = (element) => {
-    if (!element) return [];
-    if (element.childNodes.length === 0) return [element.textContent];
-    let content = [];
-    element.childNodes.forEach(child => {
-      const childTexts = getText(child);
-      content = content.concat(childTexts);
-    });
-    return content;
-  };
+  // const getText = (element) => {
+  //   if (!element) return [];
+  //   if (element.childNodes.length === 0) return [element.textContent];
+  //   let content = [];
+  //   element.childNodes.forEach(child => {
+  //     const childTexts = getText(child);
+  //     content = content.concat(childTexts);
+  //   });
+  //   return content;
+  // };
   
-  const getImages = (element) => {
-    if (!element) return [];
-    let images = [];
-    element.childNodes.forEach(child => {
-      if (child.nodeType === 1) {
-        const imageContent = handleImage(child);
-        if (imageContent.length) {
-          images.push(...imageContent);
-        } else {
-          const childImages = getImages(child);
-          images = images.concat(childImages);
-        }
-      }
-    });
-    return images;
-  };
+  // const getImages = (element) => {
+  //   if (!element) return [];
+  //   let images = [];
+  //   element.childNodes.forEach(child => {
+  //     if (child.nodeType === 1) {
+  //       const imageContent = handleImage(child);
+  //       if (imageContent.length) {
+  //         images.push(...imageContent);
+  //       } else {
+  //         const childImages = getImages(child);
+  //         images = images.concat(childImages);
+  //       }
+  //     }
+  //   });
+  //   return images;
+  // };
 
   // const extractFileName = (url) => {
   //   const regexSelectImgFileName = /([^\/]+)$/;
@@ -163,6 +163,5 @@ function extractDataFromTable() {
   return results;
 
 }
-extractDataFromTable();
 
 export { extractDataFromTable };
